+++ CHANGES.md
@@ -2,6 +2,11 @@
 
 ## Ongoing
 
+- avoid displaying detailed pytest-benchmarks statistics
+- unify testing of reasoning modes with testing of base solving
+- replace `.expected.models` with `.expected.stats` contains key-value pairs
+- use pytests's xfail and skip
+
 ## v0.0.2.dev13
 
 - introduce caching for Clorm type resolution within `myClorm.py`.